requirements.txt
# Local development (WSL/Linux)
# Python 3.10+

# Data processing
datasets>=2.18.0
pandas>=2.0.0
scikit-learn>=1.4.0

# Hugging Face core
transformers>=4.40.0
tokenizers>=0.19.0
accelerate>=0.29.0
huggingface_hub>=0.22.0

# Fine-tuning (HF Jobs scripts use PEP 723 inline deps, these are for local dev)
peft>=0.14.0
bitsandbytes>=0.43.0  # QLoRA 4-bit quantisation
trl>=0.12.0            # SFTTrainer for QLoRA training

# Analysis & plotting
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
